moved legion and mpilegion related tests into their respective subdirs
