+++ flecsi/execution/test/legion/connectivity.cc
