+++ flecsi/execution/test/legion/connectivity_partition.cc
