+++ flecsi/execution/test/mpilegion/simple2d-8x8.msh
